+++ source/video_trim2/changelog.md
@@ -1,4 +1,6 @@
 
+**<span style="color:#56adda">0.0.3</span>**
+- add debug logger output
 
 **<span style="color:#56adda">0.0.2</span>**
 - revise '-to' time calculation